# area(DCache): reduce 8 way to 4 way